Move Wrapping<T> definition to wrapping mod
